requirements.txt
dj-database-url==0.4.2
Django==1.11.2
django-api-commons==1.0.2
django-cors-headers==2.1.0
djangorestframework==3.5.4
gunicorn==19.7.1
psycopg2==2.7.1
pytz==2017.2
typing==3.5.3.0
